test: move pure value-type tests into a parallel suite
Extract the 16 self-contained tests (CappedDataCollector, DownsampledImage-
Sizing, RemoteImageURLPolicy, RelayURLValidator) out of the serialized
whitenoise_macTests suite into a new PureValueTests suite. These touch only
production value types — no UserDefaults.standard, no RemoteImageURLProtocol-
Stub statics, no shared fixtures — so they are safe to run in parallel.
Swift Testing runs separate suites concurrently, so they no longer wait
behind the serialized suite.

All 208 tests pass (now across 2 suites).
